second commit

+++ main.go
@@ -68,7 +68,7 @@ func (s ParcelService) PrintClientParcels(client int) error {
 }
 
 func (s ParcelService) NextStatus(number int) error {
-	parcel, err := s.store.Get(int64(number))
+	parcel, err := s.store.Get(number)
 	if err != nil {
 		return err
 	}
@@ -85,14 +85,14 @@ func (s ParcelService) NextStatus(number int) error {
 
 	fmt.Printf("У посылки № %d новый статус: %s\n", number, nextStatus)
 
-	return s.store.SetStatus(int64(number), nextStatus)
+	return s.store.SetStatus(number, nextStatus)
 }
 
-func (s ParcelService) ChangeAddress(number int64, address string) error {
+func (s ParcelService) ChangeAddress(number int, address string) error {
 	return s.store.SetAddress(number, address)
 }
 
-func (s ParcelService) Delete(number int64) error {
+func (s ParcelService) Delete(number int) error {
 	return s.store.Delete(number)
 }
 
@@ -119,7 +119,7 @@ func main() {
 
 	// изменение адреса
 	newAddress := "Саратов, д. Верхние Зори, ул. Козлова, д. 25"
-	err = service.ChangeAddress(int64(p.Number), newAddress)
+	err = service.ChangeAddress(p.Number, newAddress)
 	if err != nil {
 		fmt.Println(err)
 		return
@@ -140,7 +140,7 @@ func main() {
 	}
 
 	// попытка удаления отправленной посылки
-	err = service.Delete(int64(p.Number))
+	err = service.Delete(p.Number)
 	if err != nil {
 		fmt.Println(err)
 		return
@@ -162,7 +162,7 @@ func main() {
 	}
 
 	// удаление новой посылки
-	err = service.Delete(int64(p.Number))
+	err = service.Delete(p.Number)
 	if err != nil {
 		fmt.Println(err)
 		return
